This analysis will help you to take data-based decision that will save time.
In this project we are going to know the data analysis while purchasing a home in real world scenario. If someone have specific budget and want to buy a home with certain amenities, then what should be the strategy to follow and get the best result out of this analysis.